feat(tablepanel): update TabsTablePanel and TablePanel ts
